fixed minor ui bugs

+++ app/Http/Requests/IngredientRequest.php
@@ -20,7 +20,7 @@ class IngredientRequest extends FormRequest
             'buying_price'    => ['required', 'numeric', 'min:0'],
             'unit'            => ['required', 'string', 'max:255'],
 //            'quantity'        => ['required', 'integer', 'min:0'],
-            'quantity_alert'  => ['required', 'integer', 'min:0'],
+            'quantity_alert'  => ['required', 'numeric', 'min:0'],
         ];
     }
 }
